API/requirements.txt
fastapi
uvicorn
python-dotenv
mailjet-rest
python-multipart
prometheus_client>=0.15.0